Move Constants#BOOTSTRAP_PACKAGE_PREFIXES to bootstrap so we can avoid bootstrap helper-injection in ClassloadingInstrumentation
