+++ dd-java-agent/agent-bootstrap/src/main/java/datadog/trace/bootstrap/Constants.java
@@ -1,0 +1,27 @@
+package datadog.trace.bootstrap;
+
+/**
+ * Some useful constants.
+ *
+ * <p>Idea here is to keep this class safe to inject into client's class loader.
+ */
+public final class Constants {
+
+  /**
+   * packages which will be loaded on the bootstrap classloader
+   *
+   * <p>Updates should be mirrored in
+   * datadog.trace.agent.test.SpockRunner#BOOTSTRAP_PACKAGE_PREFIXES_COPY
+   */
+  public static final String[] BOOTSTRAP_PACKAGE_PREFIXES = {
+    "datadog.slf4j",
+    "datadog.trace.api",
+    "datadog.trace.bootstrap",
+    "datadog.trace.context",
+    "datadog.trace.instrumentation.api",
+    "datadog.trace.logging",
+    "datadog.trace.util",
+  };
+
+  private Constants() {}
+}
