+++ dd-java-agent/agent-tooling/src/main/java/datadog/trace/agent/tooling/context/ContextStoreUtils.java
@@ -74,9 +74,9 @@ final class ContextStoreUtils {
   }
 
   /**
-   * Note: the value here has to be inside on of the prefixes in
-   * datadog.trace.agent.tooling.Constants#BOOTSTRAP_PACKAGE_PREFIXES. This ensures that 'isolating'
-   * (or 'module') classloaders like jboss and osgi see injected classes. This works because we
+   * Note: the value here has to be inside one of the prefixes in
+   * datadog.trace.bootstrap.Constants#BOOTSTRAP_PACKAGE_PREFIXES. This ensures that 'isolating' (or
+   * 'module') classloaders like jboss and osgi see injected classes. This works because we
    * instrument those classloaders to load everything inside bootstrap packages.
    */
   private static final String DYNAMIC_CLASSES_PACKAGE =
